+++ lua/me/plugins/nvim-cmp.lua
@@ -2,8 +2,8 @@ return {
 	"hrsh7th/nvim-cmp",
 	event = "InsertEnter",
 	dependencies = {
-		-- Source for text in buffer
-		"hrsh7th/cmp-buffer",
+		"hrsh7th/cmp-buffer", -- Source for text in buffer
+		"hrsh7th/cmp-path", -- Source for file paths
 	},
 	config = function()
 		local cmp = require("cmp")
@@ -39,6 +39,8 @@ return {
 			sources = cmp.config.sources({
 				-- Text within current buffer
 				{ name = "buffer" },
+				-- File paths
+				{ name = "path", option = { trailing_slash = true, include_hidden = true } },
 			}),
 		})
 	end,